// src/src/app/shared/models/generic/ResumeStrategie.ts
/**
 * Opérations liées aux stratégies
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * un résumé de stratégie
 */
export interface ResumeStrategie extends models.Objet {
    /**
     * priorité de la stratégie
     */
    priorite?: number;

    /**
     * code du modèle 
     */
    codeModele?: string;

    /**
     * code de l'état d'activation de la strategie
     */
    codeEtat?: string;

}