// src/src/app/shared/models/generic/Strategie.ts
/**
 * Opérations liées aux stratégies
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * une stratégie
 */
export interface Strategie extends models.Objet {
    /**
     * priorité de la stratégie
     */
    priorite?: number;

    /**
     * code du modèle 
     */
    codeModele?: string;

    /**
     * code de l'état d'activation de la strategie
     */
    codeEtat?: string;

    /**
     * Map codeInfo equipement -> message à appliquer sur l'équipement pour cette stratégies
     */
    messages?: { [key: string]: models.MessageStrategie; };

}